requirements.txt
tweepy==4.14.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
psutil==5.9.5
aiohttp==3.8.5